Make lots of changes
Yea I changed a whole bunch of things. I should have been commiting but
I wasn't. You can run the program now though, and mess around with
variables. I still want to make it do more. And I need to ensure it's
working properly and not just not crashing or giving an error.