Add code for a swing swing_tester
I started working on this and realized that the contract values were
most likely taken at a different date. So when I'm simulating that its
one date the contracts values are captured from a different date.
Because of that, I'm going to hold off for a bit until I get some
clarification from the EODHDAPIs support.