Add different swing tester
This one test based on the number of days. It's just checking to see how
often you'd be correct guessing when a swing will happen. I need to
update it so it doesn't start counting until an actual swing.

Also, I found out that Python gets confused when things are in different
folders. I created new files and moved them into swing_folder.