Add ability to estimate profit and loss
I used the delta to make an assumption on how much money would be won or
loss. If the delta is 0.30, you can act as if you own 30 stocks * how
ever many contracts you have. Your upside is infinite, and your downside
is whatever your premium is. It will also start to buy more contracts
as the balance goes up. Interesting enough, the end balance is lower
once you start to buy more. It's because once you get over $2000, it
starts to buy 2. And the first contract it buys after that it ends up
losing money on.