Modify the is_swing function to account for gaps
I went back to manually check the last date it checked. The candle was
red, and the next day was green, but it still closed below the current
days candle. That's because there was a large gap down overnight. So,
I changed the is_swing function to account for that. Now there has to be
an opposite candle, and it has to close higher or lower, depending on
what the current day candle is. It lowered the percentage to 55%, but
that's still an edge, albeit not by much. The next 4 days that it
forecasted would have all been correct though.