Add code to handle breakeven prices
So, I realized that it wasn't taking into account breakeven prices. I
added this by assuming you'd get a contract that is just in the money.
And then add whatever the contract price is. At the moment, you lose
money and can only trade a few times until you run out. This is because
the contract cost are high which mean your potential losses are high.
Tomorrow I want to check the prices of contracts that expire the same
day.