Add function to find trend trade chance
The idea is simple. Look back n_days to determine the trend, and then
guess what the next day is based on the trend. Go with the trend. It
seems that just going based on the previous day will give you the best
odds, however, it's not much higher than 50%.