Add trade function to trade based on chance
Unless I'm doing something incorrectly, which is possible, it basically
seems like a 50% chance regardless of what you do. It's trying to use
probability based on frequency data that is collected based on each days
candle. For example, out of 393 counts, 211 of those will be 1 day.
Meaning, there is a 54% of the stock flipping every 1 day or a 46% of it
staying the same direction for another day. Once it gets to 2 days,
those odds will obviously change.

I need to ensure its counting the days properly though.